Hi there!
Due to a fair amount of requests to implement a custom version of CSS3 Memory for other people, and a request to use the artwork, as well as a miscellany of e-mails, I decided to open-source CSS3 Memory.

I haven't the slightest clue on how this open-source thing works, but hope to pick up stuff along the way.

Have fun with it! And I'd love for you to keep me in the loop if you are going to do something awesome.